create a ClusterVNodeOptions validtor for EventStoreDB start up validations
